chore: extract mocking functions to uikit-testing-tools package
