+++ packages/uikit-chat-hooks/src/__tests__/handler/useUserEventHandler.test.ts
@@ -1,13 +1,13 @@
 import { renderHook } from '@testing-library/react-native';
 
 import { UserEventHandler } from '@sendbird/chat';
+import { createMockSendbirdChat } from '@sendbird/uikit-testing-tools';
 
 import { useUserEventHandler } from '../../handler/useUserEventHandler';
-import { createMockSendbird } from '../__mocks__/createMockSendbirdSDK';
 
 describe('useUserEventHandler', () => {
   it('should call addUserEventHandler with the correct arguments', () => {
-    const sdk = createMockSendbird();
+    const sdk = createMockSendbirdChat();
     const handlerId = 'test_handler_id';
 
     renderHook(() => useUserEventHandler(sdk, handlerId, {}));
@@ -16,7 +16,7 @@ describe('useUserEventHandler', () => {
   });
 
   it('should call removeUserEventHandler with the correct arguments', () => {
-    const sdk = createMockSendbird();
+    const sdk = createMockSendbirdChat();
     const handlerId = 'test_handler_id';
 
     const { unmount } = renderHook(() => useUserEventHandler(sdk, handlerId, {}));
@@ -27,7 +27,7 @@ describe('useUserEventHandler', () => {
   });
 
   it('should connection handler triggered when event received', () => {
-    const sdk = createMockSendbird();
+    const sdk = createMockSendbirdChat();
     const handlerId = 'test_handler_id';
     const hookHandler = { onTotalUnreadMessageCountUpdated: jest.fn() };
     const eventArgs = [123, {}];
