+++ packages/uikit-testing-tools/README.md
@@ -1,0 +1,9 @@
+# @sendbird/uikit-utils
+
+Utility library for Sendbird UIKit
+
+## Installation
+
+```sh
+npm i @sendbird/uikit-utils
+```
