+++ packages/uikit-testing-tools/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "@sendbird/uikit-chat-hooks",
-  "version": "2.4.1",
-  "description": "React hooks that built with SendbirdChat SDK",
+  "name": "@sendbird/uikit-testing-tools",
+  "version": "2.4.0",
+  "private": true,
+  "description": "UIKit testing tools",
   "main": "lib/commonjs/index",
   "module": "lib/module/index",
   "types": "lib/typescript/src/index.d.ts",
@@ -25,7 +26,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/sendbird/sendbird-uikit-react-native.git",
-    "directory": "packages/uikit-chat-hooks"
+    "directory": "packages/uikit-testing-tools"
   },
   "author": "Sendbird <support@sendbird.com>",
   "license": "SEE LICENSE IN LICENSE.md",
@@ -37,19 +38,17 @@
     "registry": "https://registry.npmjs.org/",
     "access": "public"
   },
-  "dependencies": {
-    "@sendbird/uikit-utils": "2.4.1"
-  },
   "devDependencies": {
-    "@sendbird/uikit-testing-tools": "2.4.0",
+    "@sendbird/chat": "^4.3.0",
+    "@sendbird/uikit-utils": "2.4.0",
+    "@types/jest": "^29.4.0",
     "@types/react": "*",
+    "@types/react-native": "*",
+    "jest": "^29.4.3",
     "react": "17.0.2",
+    "react-native": "0.67.5",
     "react-native-builder-bob": "^0.18.0",
     "typescript": "4.9.4"
-  },
-  "peerDependencies": {
-    "@sendbird/chat": "^4.3.0",
-    "react": ">=16.13.1"
   },
   "react-native-builder-bob": {
     "source": "src",
@@ -64,10 +63,5 @@
         }
       ]
     ]
-  },
-  "typedoc": {
-    "entryPoint": "./src/index.ts",
-    "readmeFile": "./README.md",
-    "displayName": "@sendbird/uikit-chat-hooks"
   }
 }
