+++ packages/uikit-testing-tools/tsconfig.json
@@ -3,9 +3,6 @@
   "references": [
     {
       "path": "../uikit-utils"
-    },
-    {
-      "path": "../uikit-testing-tools"
     }
   ],
   "compilerOptions": {
